refactor package algorithm
