+++ pkg/algorithm/hived_algorithm.go
@@ -30,7 +30,6 @@ import (
 	core "k8s.io/api/core/v1"
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/klog"
-	"math"
 	"math/rand"
 	"sync"
 )
@@ -758,7 +757,7 @@ func (h *HivedAlgorithm) scheduleGuaranteedAffinityGroup(
 					// allocate a new physical cell to the preassigned cell. input a copy of the free cell list
 					// because during the scheduling we should not make in-place change to the data structures
 					// (they can be modified only when adding or deleting pods)
-					c := buddyAlloc(h.getTmpFreeCellList(sr.chain), pac.GetLevel(), suggestedNodes)
+					c := buddyAlloc(h.freeCellList[sr.chain].copy(), pac.GetLevel(), suggestedNodes)
 					if c == nil {
 						panic(fmt.Sprintf(
 							"VC Safety Broken: Cannot find physical cell for a VC cell: %v", pac.GetAddress()))
@@ -793,16 +792,6 @@ func (h *HivedAlgorithm) scheduleOpportunisticAffinityGroup(
 			sr.affinityGroupPodNums, sr.priority, sr.chain)
 	}
 	return physicalPlacement
-}
-
-// getTmpFreeCellList returns a copy of the free cell list.
-func (h *HivedAlgorithm) getTmpFreeCellList(chain CellChain) ChainCellList {
-	ccl := ChainCellList{}
-	for l := CellLevel(1); l <= CellLevel(len(h.freeCellList[chain])); l++ {
-		ccl[l] = make(CellList, len(h.freeCellList[chain][l]))
-		copy(ccl[l], h.freeCellList[chain][l])
-	}
-	return ccl
 }
 
 // createAllocatedAffinityGroup creates a new affinity group, and confirms the allocated resources.
@@ -1254,446 +1243,4 @@ func generatePodScheduleResult(
 			},
 		}
 	}
-}
-
-// generateAffinityGroupBindInfo writes the physical and virtual placements of an affinity group
-// into a a series of AffinityGroupMemberBindInfos, and returns the allocated node and GPU addresses
-// of the current pod.
-func generateAffinityGroupBindInfo(
-	groupPhysicalPlacement groupPhysicalPlacement,
-	groupVirtualPlacement groupVirtualPlacement,
-	cellLevelToType map[CellChain]map[CellLevel]api.CellType,
-	currentGpuNum int32,
-	currentPodIndex int32,
-	group *AlgoAffinityGroup,
-	groupName string,
-	suggestedNodes common.Set) (
-	affinityGroupBindInfo []api.AffinityGroupMemberBindInfo,
-	nodesNotInSuggested []string,
-	selectedNode string,
-	selectedGpuIndices []int32,
-	chain string) {
-
-	if groupPhysicalPlacement == nil {
-		return
-	}
-	affinityGroupBindInfo = make([]api.AffinityGroupMemberBindInfo, len(groupPhysicalPlacement))
-	groupMemberIndex := 0
-	for podGpuNum, podPhysicalPlacements := range groupPhysicalPlacement {
-		mbi := api.AffinityGroupMemberBindInfo{
-			PodPlacements: make([]api.PodPlacementInfo, len(podPhysicalPlacements)),
-		}
-		for podIndex := int32(0); podIndex < int32(len(podPhysicalPlacements)); podIndex++ {
-			mbi.PodPlacements[podIndex].PhysicalGpuIndices = make([]int32, podGpuNum)
-			mbi.PodPlacements[podIndex].PreassignedCellTypes = make([]api.CellType, podGpuNum)
-			for gpuIndex := int32(0); gpuIndex < podGpuNum; gpuIndex++ {
-				pGpu := podPhysicalPlacements[podIndex][gpuIndex]
-				if pGpu == nil {
-					if group == nil {
-						panic(fmt.Sprintf("The first pod in group %v was allocated invalid resource", groupName))
-					}
-					// if the physical placement of this pod is not found (e.g., removed due to reconfiguration),
-					// we will insist the decision by retrieving it from other pods
-					mbi.PodPlacements[podIndex], chain = retrieveMissingPodPlacement(group, podGpuNum, podIndex)
-					klog.Warningf(
-						"pod placement has been invalid and is retrieved from annotation of other pods: node %v, GPU %v",
-						mbi.PodPlacements[podIndex].PhysicalNode, mbi.PodPlacements[podIndex].PhysicalGpuIndices[gpuIndex])
-				} else {
-					nodes, gpuIndices := pGpu.(*PhysicalCell).GetPhysicalPlacement()
-					// here each cell (i.e., pGpu) is only one GPU, hence we takes the first element
-					// in its "nodes" and "gpuIndices" as the node and GPU address
-					if mbi.PodPlacements[podIndex].PhysicalNode == "" {
-						mbi.PodPlacements[podIndex].PhysicalNode = nodes[0]
-						if !suggestedNodes.Contains(nodes[0]) {
-							nodesNotInSuggested = append(nodesNotInSuggested, nodes[0])
-						}
-					}
-					mbi.PodPlacements[podIndex].PhysicalGpuIndices[gpuIndex] = gpuIndices[0]
-					if groupVirtualPlacement != nil {
-						vGpu := groupVirtualPlacement[podGpuNum][podIndex][gpuIndex].(*VirtualCell)
-						mbi.PodPlacements[podIndex].PreassignedCellTypes[gpuIndex] =
-							cellLevelToType[vGpu.GetChain()][vGpu.GetPreAssignedCell().GetLevel()]
-					} else {
-						mbi.PodPlacements[podIndex].PreassignedCellTypes[gpuIndex] = ""
-					}
-				}
-			}
-		}
-		if podGpuNum == currentGpuNum {
-			selectedNode = mbi.PodPlacements[currentPodIndex].PhysicalNode
-			selectedGpuIndices = mbi.PodPlacements[currentPodIndex].PhysicalGpuIndices
-			if pGpu := groupPhysicalPlacement[currentGpuNum][currentPodIndex][0]; pGpu != nil {
-				chain = string(pGpu.GetChain())
-			}
-		}
-		affinityGroupBindInfo[groupMemberIndex] = mbi
-		groupMemberIndex++
-	}
-	return affinityGroupBindInfo, nodesNotInSuggested, selectedNode, selectedGpuIndices, chain
-}
-
-// collectPreemptionVictims collects preemption victims of an affinity group.
-// If any of the GPUs allocated for the whole group is still used by a pod,
-// we will wait for the preemption, as a group is gang-scheduled.
-func collectPreemptionVictims(
-	groupPhysicalPlacement groupPhysicalPlacement,
-	priority CellPriority,
-	groupName string) (map[string]common.Set, []string) {
-
-	preemptionVictims := map[string]common.Set{}
-	var nodesHaveVictims []string
-	for gpuNum := range groupPhysicalPlacement {
-		for podIndex := range groupPhysicalPlacement[gpuNum] {
-			for _, gpu := range groupPhysicalPlacement[gpuNum][podIndex] {
-				if gpu == nil {
-					continue
-				}
-				pGpu := gpu.(*PhysicalCell)
-				if victimGroup := pGpu.GetAffinityGroup(); victimGroup != nil && victimGroup.name != groupName {
-					// There are two cases of finding a running pod on the allocated resources:
-					// 1. the running pod is a preemption victim.
-					// 2. the running pod used resource partially released by the current group,
-					// but the group wants to schedule a pod again.
-					// Our principle is we allow preemption if the running pod's priority is lower than that
-					// of the group to be scheduled (the 2nd case may also satisfy this condition, and we
-					// allow such preemption). Otherwise the running pod cannot be preempted, and the pod
-					// to be scheduled will wait.
-					if pGpu.GetPriority() >= priority {
-						panic(fmt.Sprintf(
-							"Resources previously allocated (%v) has been allocated to "+
-								"another non-preemptible group %v; pod should wait",
-							pGpu.GetPhysicalPlacementString(), victimGroup.name))
-					}
-					// for any victim pod, gang-preempt all the other pods from the same affinity group
-					for _, victims := range victimGroup.allocatedPods {
-						for _, v := range victims {
-							if v != nil {
-								if _, ok := preemptionVictims[v.Spec.NodeName]; !ok {
-									preemptionVictims[v.Spec.NodeName] = common.NewSet()
-									nodesHaveVictims = append(nodesHaveVictims, v.Spec.NodeName)
-								}
-								preemptionVictims[v.Spec.NodeName].Add(v)
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-	return preemptionVictims, nodesHaveVictims
-}
-
-// retrieveMissingPodPlacement finds the placement of a pod from the annotation of other pods in the same group
-// when the pod's placement has been invalid (i.e., not found in the spec).
-func retrieveMissingPodPlacement(group *AlgoAffinityGroup, gpuNum int32, podIndex int32) (api.PodPlacementInfo, string) {
-	for _, pods := range group.allocatedPods {
-		for _, p := range pods {
-			if p != nil {
-				info := internal.ExtractPodBindInfo(p)
-				for _, mbi := range info.AffinityGroupBindInfo {
-					if gpuNum == int32(len(mbi.PodPlacements[0].PhysicalGpuIndices)) {
-						return mbi.PodPlacements[podIndex], info.CellChain
-					}
-				}
-			}
-		}
-	}
-	panic(fmt.Sprintf(
-		"No allocated pod found in an allocated group %v when retrieving placement for pod %v with GPU number %v", group.name, podIndex, gpuNum))
-}
-
-// buddyAlloc allocates a free cell at a certain level from a free list.
-// It splits a higher-level cell when there is no free cell at the current level.
-// As the input cell list is a copy of the real free list and hence is one-off,
-// we won't remove a returned cell from it.
-func buddyAlloc(freeList ChainCellList, level CellLevel, suggestedNodes common.Set) *PhysicalCell {
-	if len(freeList[level]) == 0 && level < CellLevel(len(freeList)) {
-		higherCell := buddyAlloc(freeList, level+1, suggestedNodes)
-		if higherCell != nil {
-			freeList[level] = append(freeList[level], higherCell.GetChildren()...)
-		}
-	}
-	if len(freeList[level]) == 0 {
-		return nil
-	}
-	return getFewestOpporPhysicalCell(freeList[level], suggestedNodes)
-}
-
-// getFewestOpporPhysicalCell selects a physical cell with the minimum number of opportunistic pods from a cell list.
-// This function will try to avoid using cells whose resources are not within the suggested nodes if possible.
-// In case there is no cell within the suggested nodes, it will try to randomly return a healthy node with at least
-// one opportunistic pod to try preemption. If that also fails, it just returns the min-opportunistic cell,
-// regardless of if it is suggested or healthy.
-func getFewestOpporPhysicalCell(cl CellList, suggestedNodes common.Set) *PhysicalCell {
-	fewestOpporNum := int32(math.MaxInt32)
-	fewestOpporNumSuggested := int32(math.MaxInt32)
-	var fewestOpporCell *PhysicalCell
-	var fewestOpporCellSuggested *PhysicalCell
-	var preemptibleCells []*PhysicalCell
-	for _, c := range cl {
-		if pc := c.(*PhysicalCell); pc.GetVirtualCell() == nil && pc.GetPreBoundVirtualCell() == nil {
-			opporNum := pc.GetUsedGpuNumAtPriorities()[opportunisticPriority]
-			if opporNum < fewestOpporNum {
-				fewestOpporNum = opporNum
-				fewestOpporCell = pc
-			}
-			allNodesInSuggested := true
-			nodes, _ := pc.GetPhysicalPlacement()
-			for _, n := range nodes {
-				if !suggestedNodes.Contains(n) {
-					allNodesInSuggested = false
-					break
-				}
-			}
-			if allNodesInSuggested && opporNum < fewestOpporNumSuggested {
-				fewestOpporNumSuggested = opporNum
-				fewestOpporCellSuggested = pc
-			}
-			if opporNum > 0 {
-				preemptibleCells = append(preemptibleCells, pc)
-			}
-		}
-	}
-	var selectedCell *PhysicalCell
-	if fewestOpporCellSuggested != nil {
-		selectedCell = fewestOpporCellSuggested
-		nodes, _ := selectedCell.GetPhysicalPlacement()
-		klog.Infof("Selected a cell within suggested nodes: %v, nodes %v", selectedCell.GetAddress(), common.ToJson(nodes))
-	} else if len(preemptibleCells) > 0 {
-		// If we cannot find a cell within suggested nodes, we will try to preempt some pods instead of
-		// directly returning the fewestOpporCell (because this cell could be a bad node, we should not return it).
-		// Also, we will choose a random cell, to avoid always returning the same cell (similar to above,
-		// if we always return the same cell, it might be a bad node, preempting pods on a bad node won't bring
-		// it back to the suggested nodes)
-		selectedCell = preemptibleCells[rand.Int31n(int32(len(preemptibleCells)))]
-		nodes, _ := selectedCell.GetPhysicalPlacement()
-		klog.Infof("Selected a cell not within suggested nodes but preempting opportunistic pods may help: %v, nodes %v",
-			selectedCell.GetAddress(), common.ToJson(nodes))
-	} else if fewestOpporCell == nil {
-		panic("VC Safety Broken: Cannot find any physical cell that has not been bound to a virtual cell")
-	} else {
-		selectedCell = fewestOpporCell
-		nodes, _ := selectedCell.GetPhysicalPlacement()
-		klog.Infof("Selected a cell not within suggested nodes and no preemption can help: %v, nodes %v",
-			selectedCell.GetAddress(), common.ToJson(nodes))
-	}
-	return selectedCell
-}
-
-// mapNonPreassignedCellToPhysical maps a virtual cell (possibly inside a preassigned one) to
-// one of the cell inside the physical cell of the preassigned cell. This operation keeps the
-// inner-cell topology equivalent, by recursively binding the cells inside the preassigned one.
-func mapNonPreassignedCellToPhysical(c *VirtualCell, suggestedNodes common.Set) *PhysicalCell {
-	if c.GetPhysicalCell() != nil {
-		return c.GetPhysicalCell()
-	} else if c.GetPreBoundPhysicalCell() != nil {
-		return c.GetPreBoundPhysicalCell()
-	} else {
-		parentPhysical := mapNonPreassignedCellToPhysical(c.GetParent().(*VirtualCell), suggestedNodes)
-		pc := getFewestOpporPhysicalCell(parentPhysical.GetChildren(), suggestedNodes)
-		if pc == nil || pc.GetPriority() > opportunisticPriority {
-			panic(fmt.Sprintf("VC Safety Broken: Cannot find physical cell for %v", c.GetAddress()))
-		}
-		c.SetPreBoundPhysicalCell(pc)
-		pc.SetPreBoundVirtualCell(c)
-		return pc
-	}
-}
-
-// clearPreBindings clears the temporary bindings created during scheduling.
-func clearPreBindings(virtualPlacement groupVirtualPlacement) {
-	for _, podPlacements := range virtualPlacement {
-		for _, podGpus := range podPlacements {
-			for _, gpu := range podGpus {
-				for gpu != nil {
-					vGpu := gpu.(*VirtualCell)
-					if pGpu := vGpu.GetPreBoundPhysicalCell(); pGpu != nil {
-						pGpu.SetPreBoundVirtualCell(nil)
-						vGpu.SetPreBoundPhysicalCell(nil)
-						gpu = gpu.GetParent()
-					} else {
-						break
-					}
-				}
-			}
-		}
-	}
-}
-
-// getPodIndex finds the index of a pod in its group according to its placement.
-func getPodIndex(podPlacements []api.PodPlacementInfo, node string, gpuIndex int32) int32 {
-	for podIndex, placement := range podPlacements {
-		if placement.PhysicalNode == node && common.Int32SliceContains(placement.PhysicalGpuIndices, gpuIndex) {
-			return int32(podIndex)
-		}
-	}
-	return -1
-}
-
-// mapNonPreassignedCellToVirtual maps a physical cell (possibly allocated to a non-preassigned virtual cell)
-// to the corresponding virtual cell. This can be viewed as an inverse operation of mapNonPreassignedCellToPhysical,
-// used for finding the virtual cell when adding an allocated pod.
-func mapNonPreassignedCellToVirtual(
-	c *PhysicalCell,
-	vccl ChainCellList,
-	preassignedLevel CellLevel,
-	p CellPriority) (*VirtualCell, string) {
-
-	if c.GetVirtualCell() != nil {
-		return c.GetVirtualCell(), ""
-	} else if c.GetLevel() == preassignedLevel {
-		if preassignedVirtual := getLowestPriorityCell(vccl[preassignedLevel], p); preassignedVirtual == nil {
-			return nil, fmt.Sprintf("insufficient free cell in the VC at the preassigned level (%v)", preassignedLevel)
-		} else {
-			return preassignedVirtual.(*VirtualCell), ""
-		}
-	} else if c.GetParent() == nil {
-		return nil, fmt.Sprintf(
-			"physical and virtual cell hierarchies not match (cannot reach the preassigned level %v in physical)",
-			preassignedLevel)
-	} else {
-		parentVirtual, message := mapNonPreassignedCellToVirtual(c.GetParent().(*PhysicalCell), vccl, preassignedLevel, p)
-		if parentVirtual == nil {
-			return nil, message
-		} else {
-			return getLowestPriorityCell(parentVirtual.GetChildren(), p).(*VirtualCell), ""
-		}
-	}
-}
-
-// getLowestPriorityCell returns a cell with the lowest priority among the cells
-// whose priorities are lower than the given priority (p).
-func getLowestPriorityCell(cl CellList, p CellPriority) Cell {
-	lowestPriority := maxGuaranteedPriority
-	var lowestPriorityCell Cell
-	for _, c := range cl {
-		pp := c.GetPriority()
-		if pp == freePriority {
-			return c
-		} else if pp < p && pp < lowestPriority {
-			lowestPriority = pp
-			lowestPriorityCell = c
-		}
-	}
-	return lowestPriorityCell
-}
-
-// bindCell binds a virtual cell to a physical cell and its parent recursively.
-func bindCell(pc *PhysicalCell, vc *VirtualCell) {
-	for vc.GetPhysicalCell() == nil {
-		pc.SetVirtualCell(vc)
-		vc.SetPhysicalCell(pc)
-		message := fmt.Sprintf("Cells bound: %v and %v", vc.GetAddress(), pc.GetAddress())
-		if pc.IsReserved() {
-			message += " (reservation)"
-		}
-		klog.Info(message)
-		if vc.GetParent() == nil {
-			break
-		}
-		vc = vc.GetParent().(*VirtualCell)
-		pc = pc.GetParent().(*PhysicalCell)
-	}
-}
-
-// unbindCell unbinds a virtual cell with a physical cell and its parent recursively.
-func unbindCell(c *PhysicalCell) {
-	boundVirtual := c.GetVirtualCell()
-	for !boundVirtual.GetPhysicalCell().IsReserved() {
-		boundPhysical := boundVirtual.GetPhysicalCell()
-		klog.Infof("Cells unbound: %v and %v", boundVirtual.GetAddress(), boundPhysical.GetAddress())
-		boundVirtual.SetPhysicalCell(nil)
-		boundPhysical.SetVirtualCell(nil)
-		if boundVirtual.GetParent() == nil {
-			break
-		} else {
-			unbindParent := true
-			for _, cc := range boundVirtual.GetParent().GetChildren() {
-				if child := cc.(*VirtualCell); child.GetPhysicalCell() != nil {
-					unbindParent = false
-					break
-				}
-			}
-			if !unbindParent {
-				break
-			}
-			boundVirtual = boundVirtual.GetParent().(*VirtualCell)
-		}
-	}
-}
-
-// setPriority sets priority and state for a cell and its parent recursively, guaranteeing that
-// (i) the priority of a cell is the max of those of its children.
-// (ii) a cell is in "Used" state if any of its children is "Used", otherwise "Free".
-func setPriority(c Cell, p CellPriority) {
-	originalPriority := c.GetPriority()
-	c.SetPriority(p)
-	if parent := c.GetParent(); parent != nil {
-		if p > parent.GetPriority() {
-			setPriority(parent, p)
-		} else if originalPriority == parent.GetPriority() && p < originalPriority {
-			maxBuddyPriority := freePriority
-			for _, buddy := range parent.GetChildren() {
-				if buddy.GetPriority() > maxBuddyPriority {
-					maxBuddyPriority = buddy.GetPriority()
-				}
-			}
-			setPriority(parent, maxBuddyPriority)
-		}
-	}
-}
-
-// updateUsedGpuNumAtPriority updates the number of used GPUs at a priority for a cell
-// and its parent recursively.
-func updateUsedGpuNumAtPriority(c Cell, p CellPriority, increase bool) {
-	for c != nil {
-		delta := int32(-1)
-		if increase {
-			delta = 1
-		}
-		c.IncreaseUsedGpuNumAtPriority(p, delta)
-		c = c.GetParent()
-	}
-}
-
-// allPodsReleased checks if all the pods of an affinity group were released.
-func allPodsReleased(allocatedPods map[int32][]*core.Pod) bool {
-	for _, pods := range allocatedPods {
-		for _, p := range pods {
-			if p != nil {
-				return false
-			}
-		}
-	}
-	return true
-}
-
-// inFreeCellList checks if a physical cell (or its ancestor) is in the global free cell list.
-func inFreeCellList(c *PhysicalCell) bool {
-	for {
-		if c.GetVirtualCell() != nil || c.IsSplit() {
-			return false
-		}
-		if c.GetParent() == nil || c.GetParent().(*PhysicalCell).IsSplit() {
-			return true
-		}
-		c = c.GetParent().(*PhysicalCell)
-	}
-}
-
-// generateOpporVirtualCell generates a fake virtual cell in a VC's API status
-// for an opportunistic cell used by the VC.
-func generateOpporVirtualCell(pc *api.PhysicalCellStatus) *api.VirtualCellStatus {
-	vc := &api.VirtualCellStatus{
-		CellStatus: api.CellStatus{
-			GpuType:         pc.GpuType,
-			CellType:        pc.CellType,
-			CellAddress:     pc.CellAddress + "-opp",
-			CellState:       api.CellUsed,
-			CellHealthiness: pc.CellHealthiness,
-			CellPriority:    api.OpportunisticPriority,
-		},
-		PhysicalCell: pc,
-	}
-	return vc
 }
